refactor: BurgerIcon & BurgerMenu
