+++ src/components/UI/burger/burgerMenu/styles.module.scss
@@ -9,7 +9,13 @@
   z-index: 10;
   height: 100%;
   transition: all 0.5s ease;
-  &.active {
+}
+
+.active {
+  background-color: blue;
+  position: absolute;
+  inset: 0;
+  & .menu {
     left: 0;
     transition: all 0.5s ease;
   }
