+++ src/components/UI/burger/index.tsx
@@ -1,28 +1,2 @@
-import cn from "classnames";
-import { FC, MouseEvent } from "react";
-import s from "./styles.module.scss";
-
-interface PropsType {
-  handleClick: (event: MouseEvent<HTMLDivElement>) => void;
-  open: boolean;
-}
-
-const Burger: FC<PropsType> = ({ handleClick, open }) => {
-  return (
-    <>
-      <div
-        className={cn("burgerContainer", {
-          [s.active]: open,
-        })}
-        onClick={(e) => handleClick(e)}
-      >
-        <span></span>
-        <span></span>
-        <span></span>
-      </div>
-      <div className={cn(s.menu, { [s.active]: open })}></div>
-    </>
-  );
-};
-
-export default Burger;
+export { default as BurgerIcon } from "./burgerIcon";
+export { default as BurgerMenu } from "./burgerMenu";
