+++ src/components/UI/index.ts
@@ -1,5 +1,4 @@
 export { default as Nav } from "./nav";
-export { default as Burger } from "./burger";
 export { default as Slider } from "./slider";
 export { default as Button } from "./button";
 export { default as Modal } from "./modal";
