+++ src/components/layout/header/index.tsx
@@ -1,7 +1,8 @@
-import { Burger, Nav } from "@/components/UI";
+import { Nav } from "@/components/UI";
 import { FC, MouseEvent } from "react";
 import { Link } from "react-router-dom";
 import s from "./styles.module.scss";
+import { BurgerIcon, BurgerMenu } from "@/components/UI/burger";
 
 interface Props {
   open: boolean;
@@ -16,7 +17,8 @@ const Header: FC<Props> = ({ open, handleClick }) => {
         <span className={s.subLogo}>Shelter for pets in Boston</span>
       </div>
       <Nav open={open} />
-      <Burger handleClick={handleClick} open={open} />
+      <BurgerIcon handleClick={handleClick} open={open} />
+      <BurgerMenu handleClick={handleClick} open={open} />
     </header>
   );
 };
